Implement draw_rate and remove update event
