+++ arcade/application.py
@@ -60,7 +60,8 @@ class Window(pyglet.window.Window):
     :param str title: Title (appears in title bar)
     :param bool fullscreen: Should this be full screen?
     :param bool resizable: Can the user resize the window?
-    :param float update_rate: How frequently to update the window.
+    :param float update_rate: How frequently to run the on_update event.
+    :param float draw_rate: How frequently to run the on_draw event. (this is the FPS limit)
     :param bool antialiasing: Should OpenGL's anti-aliasing be enabled?
     :param Tuple[int,int] gl_version: What OpenGL version to request. This is ``(3, 3)`` by default \
                                        and can be overridden when using more advanced OpenGL features.
@@ -93,7 +94,8 @@ class Window(pyglet.window.Window):
         center_window: bool = False,
         samples: int = 4,
         enable_polling: bool = True,
-        gl_api: str = "gl"
+        gl_api: str = "gl",
+        draw_rate: float = 1 / 60,
     ):
         # In certain environments we can't have antialiasing/MSAA enabled.
         # Detect replit environment
@@ -138,7 +140,6 @@ class Window(pyglet.window.Window):
         try:
             super().__init__(width=width, height=height, caption=title,
                              resizable=resizable, config=config, vsync=vsync, visible=visible, style=style)
-            self.register_event_type('update')
             self.register_event_type('on_update')
         except pyglet.window.NoSuchConfigException:
             raise NoOpenGLException("Unable to create an OpenGL 3.3+ context. "
@@ -149,8 +150,12 @@ class Window(pyglet.window.Window):
             except pyglet.gl.GLException:
                 LOG.warning("Warning: Anti-aliasing not supported on this computer.")
 
-        if update_rate:
-            self.set_update_rate(update_rate)
+        # We don't call the set_draw_rate function here because unlike the updates, the draw scheduling
+        # is initially set in the call to pyglet.app.run() that is done by the run() function.
+        # run() will pull this draw rate from the Window and use it. Calls to set_draw_rate only need
+        # to be done if changing it after the application has been started.
+        self._draw_rate = draw_rate
+        self.set_update_rate(update_rate)
 
         self.set_vsync(vsync)
 
@@ -255,13 +260,78 @@ class Window(pyglet.window.Window):
 
     def run(self):
         """
-        Shortcut for :py:func:`arcade.run()`.
-
-        For example::
-
-            MyWindow().run()
+        Run the main loop.
+        After the window has been set up, and the event hooks are in place, this is usually one of the last
+        commands on the main program. This is a blocking function starting pyglet's event loop
+        meaning it will start to dispatch events such as ``on_draw`` and ``on_update``.
         """
-        arcade.run()
+
+        # Used in some unit test
+        if os.environ.get('ARCADE_TEST'):       
+            self.on_update(self._update_rate)
+            self.on_draw()
+        elif self.headless:
+            # We are entering headless more an will emulate an event loop
+            import time
+            # Ensure the initial delta time is not 0 to be
+            # more in line with how a normal window works.
+            delta_time = self._draw_rate
+            last_time = time.perf_counter()
+
+            # As long as we have a context --
+            while self.context:
+                # Select active view or window
+                active = self.current_view or self
+
+                active.on_update(delta_time)
+                if self.context:
+                    active.on_draw()
+
+                # windwow could be closed in on_draw
+                if self.context:
+                    self.flip()
+
+                now = time.perf_counter()
+                delta_time, last_time = now - last_time, now
+        else:
+            import sys
+            if sys.platform != 'win32':
+                # For non windows platforms, just do pyglet run
+                pyglet.app.run(self._draw_rate)
+            else:
+                # Ok, some Windows platforms have a timer resolution > 15 ms. That can
+                # drop our FPS to 32 FPS or so. This reduces resolution so we can keep
+                # FPS up.
+                import contextlib
+                import ctypes
+                from ctypes import wintypes
+
+                winmm = ctypes.WinDLL('winmm')
+
+                class TIMECAPS(ctypes.Structure):
+                    _fields_ = (('wPeriodMin', wintypes.UINT),
+                                ('wPeriodMax', wintypes.UINT))
+
+                def _check_time_err(err, func, args):
+                    if err:
+                        raise WindowsError('%s error %d' % (func.__name__, err))
+                    return args
+
+                winmm.timeGetDevCaps.errcheck = _check_time_err
+                winmm.timeBeginPeriod.errcheck = _check_time_err
+                winmm.timeEndPeriod.errcheck = _check_time_err
+
+                @contextlib.contextmanager
+                def timer_resolution(msecs=0):
+                    caps = TIMECAPS()
+                    winmm.timeGetDevCaps(ctypes.byref(caps), ctypes.sizeof(caps))
+                    msecs = min(max(msecs, caps.wPeriodMin), caps.wPeriodMax)
+                    winmm.timeBeginPeriod(msecs)
+                    yield
+                    winmm.timeEndPeriod(msecs)
+
+                with timer_resolution(msecs=10):
+                    pyglet.app.run(self._draw_rate)
 
     def close(self):
         """ Close the Window. """
@@ -302,15 +372,6 @@ class Window(pyglet.window.Window):
         # Center the window
         self.set_location((screen_width - window_width) // 2, (screen_height - window_height) // 2)
 
-    def update(self, delta_time: float):
-        """
-        Move everything. For better consistency in naming, use ``on_update`` instead.
-
-        :param float delta_time: Time interval since the last time the function was called in seconds.
-
-        """
-        pass
-
     def on_update(self, delta_time: float):
         """
         Move everything. Perform collision checks. Do all the game logic here.
@@ -321,18 +382,31 @@ class Window(pyglet.window.Window):
         pass
 
     def _dispatch_updates(self, delta_time: float):
-        self.dispatch_event('update', delta_time)
+        """
+        Internal function that is scheduled with Pyglet's clock, this function gets run by the clock, and
+        dispatches the on_update events.
+        """
         self.dispatch_event('on_update', delta_time)
 
     def set_update_rate(self, rate: float):
         """
-        Set how often the screen should be updated.
-        For example, self.set_update_rate(1 / 60) will set the update rate to 60 fps
+        Set how often the on_update function should be dispatched.
+        For example, self.set_update_rate(1 / 60) will set the update rate to 60 times per second.
 
         :param float rate: Update frequency in seconds
         """
+        self._update_rate = rate
         pyglet.clock.unschedule(self._dispatch_updates)
         pyglet.clock.schedule_interval(self._dispatch_updates, rate)
+
+    def set_draw_rate(self, rate: float):
+        """
+        Set how often the on_draw function should be run.
+        For example, set.set_draw_rate(1 / 60) will set the draw rate to 60 frames per second.
+        """
+        self._draw_rate = rate
+        pyglet.clock.unschedule(pyglet.app.event_loop._redraw_windows)
+        pyglet.clock.schedule_interval(pyglet.app.event_loop._redraw_windows, self._draw_rate)
 
     def on_mouse_motion(self, x: int, y: int, dx: int, dy: int):
         """
@@ -881,8 +955,7 @@ def open_window(
     """
 
     global _window
-    _window = Window(width, height, window_title, resizable=resizable, update_rate=None,
-                     antialiasing=antialiasing)
+    _window = Window(width, height, window_title, resizable=resizable, antialiasing=antialiasing)
     _window.invalid = False
     return _window
 
@@ -931,10 +1004,6 @@ class View:
         :param Tuple[int, int, int, int] viewport: The viewport range to clear
         """
         self.window.clear(color, normalized, viewport)
-
-    def update(self, delta_time: float):
-        """To be overridden"""
-        pass
 
     def on_update(self, delta_time: float):
         """To be overridden"""
