+++ arcade/window_commands.py
@@ -247,85 +247,6 @@ def finish_render():
     get_window().flip_count = 0
     get_window().flip()
 
-
-def run():
-    """
-    Run the main loop.
-    After the window has been set up, and the event hooks are in place, this is usually one of the last
-    commands on the main program. This is a blocking function starting pyglet's event loop
-    meaning it will start to dispatch events such as ``on_draw`` and ``on_update``.
-    """
-    window = get_window()
-
-    # Used in some unit test
-    if os.environ.get('ARCADE_TEST'):       
-        if window:
-            window.on_update(1 / 60)
-            window.on_draw()
-    elif window.headless:
-        # We are entering headless more an will emulate an event loop
-        import time
-        # Ensure the initial delta time is not 0 to be
-        # more in line with how a normal window works.
-        delta_time = 1 / 60
-        last_time = time.perf_counter()
-
-        # As long as we have a context --
-        while window.context:
-            # Select active view or window
-            active = window.current_view or window
-
-            active.on_update(delta_time)
-            if window.context:
-                active.on_draw()
-
-            # windwow could be closed in on_draw
-            if window.context:
-                window.flip()
-
-            now = time.perf_counter()
-            delta_time, last_time = now - last_time, now
-    else:
-        import sys
-        if sys.platform != 'win32':
-            # For non windows platforms, just do pyglet run
-            pyglet.app.run()
-        else:
-            # Ok, some Windows platforms have a timer resolution > 15 ms. That can
-            # drop our FPS to 32 FPS or so. This reduces resolution so we can keep
-            # FPS up.
-            import contextlib
-            import ctypes
-            from ctypes import wintypes
-
-            winmm = ctypes.WinDLL('winmm')
-
-            class TIMECAPS(ctypes.Structure):
-                _fields_ = (('wPeriodMin', wintypes.UINT),
-                            ('wPeriodMax', wintypes.UINT))
-
-            def _check_time_err(err, func, args):
-                if err:
-                    raise WindowsError('%s error %d' % (func.__name__, err))
-                return args
-
-            winmm.timeGetDevCaps.errcheck = _check_time_err
-            winmm.timeBeginPeriod.errcheck = _check_time_err
-            winmm.timeEndPeriod.errcheck = _check_time_err
-
-            @contextlib.contextmanager
-            def timer_resolution(msecs=0):
-                caps = TIMECAPS()
-                winmm.timeGetDevCaps(ctypes.byref(caps), ctypes.sizeof(caps))
-                msecs = min(max(msecs, caps.wPeriodMin), caps.wPeriodMax)
-                winmm.timeBeginPeriod(msecs)
-                yield
-                winmm.timeEndPeriod(msecs)
-
-            with timer_resolution(msecs=10):
-                pyglet.app.run()
-
-
 def exit():
     """
     Exits the application.
